+++ src/components/nosignal24.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { IVehicle } from "../services/vehiclesApi";
+import { IVehicle } from "../Views/Home";
 import Nosignal from "../Image/nosignal.png";
 
 interface Props {
